pyproject.toml
[tool.poetry]
name = "pytorch-lightning"
version = "0.1.0"
description = ""
authors = ["Victor Bratko <bratko.viktor2004@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
poetry-core = "^2.1.3"
pydantic = "^2.11.7"
omegaconf = "^2.3.0"
pre-commit = "^4.2.0"
flake8 = "^7.2.0"
mypy = "^1.16.1"
isort = "^6.0.1"
black = "^25.1.0"
lightning = "^2.5.1.post0"
albumentations = "^2.0.8"
torchvision = "^0.22.1"
clearml = "^2.0.0"
notebook = "^7.4.3"
timm = "^1.0.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
